Run a list of analyses

Run a list of analyses

runCcAnalyses(
  connectionDetails,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable = "drug_era",
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_era",
  nestingCohortDatabaseSchema = cdmDatabaseSchema,
  nestingCohortTable = "condition_era",
  outputFolder = "./CcOutput",
  ccAnalysisList,
  exposureOutcomeNestingCohortList,
  prefetchExposureData = FALSE,
  getDbCaseDataThreads = 1,
  selectControlsThreads = 1,
  getDbExposureDataThreads = 1,
  createCaseControlDataThreads = 1,
  fitCaseControlModelThreads = 1,
  cvThreads = 1
)

Arguments

connectionDetails

An R object of type ConnectionDetails created using the function createConnectionDetails in the DatabaseConnector package.

cdmDatabaseSchema

The name of the database schema that contains the OMOP CDM instance. Requires read permissions to this database. On SQL Server, this should specify both the database and the schema, so for example 'cdm_instance.dbo'.

tempEmulationSchema

Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created.

exposureDatabaseSchema

The name of the database schema that is the location where the exposure data used to define the exposure cohorts is available. If exposureTable = DRUG_ERA, exposureDatabaseSchema is not used but assumed to be cdmSchema. Requires read permissions to this database.

exposureTable

The tablename that contains the exposure cohorts. If exposureTable <> drug_era, then expectation is exposureTable has format of COHORT table: cohort_definition_id, subject_id, cohort_start_date, cohort_end_date.

outcomeDatabaseSchema

The name of the database schema that is the location where the data used to define the outcome cohorts is available. If outcomeTable = CONDITION_ERA, outcomeDatabaseSchema is not used. Requires read permissions to this database.

outcomeTable

The tablename that contains the outcome cohorts. If outcomeTable is not CONDITION_OCCURRENCE or CONDITION_ERA, then expectation is outcomeTable has format of COHORT table: COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.

nestingCohortDatabaseSchema

The name of the database schema that is the location where the nesting cohort is defined.

nestingCohortTable

Name of the table holding the nesting cohort. This table should have the same structure as the cohort table.

outputFolder

Name of the folder where all the outputs will written to.

ccAnalysisList

A list of objects of type ccAnalysis as created using the createCcAnalysis function.

exposureOutcomeNestingCohortList

A list of objects of type exposureOutcomeNestingCohort as created using the createExposureOutcomeNestingCohort function.

prefetchExposureData

Should exposure data for the entire nesting cohort be fetched at the beginning, or should exposure data be fetch later specifically for a set of cases and controls. Prefetching can be faster when there are many outcomes but only few exposures. Prefetching does not speed up performance when covariates also need to be constructed.

getDbCaseDataThreads

The number of parallel threads to use for building the caseData objects.

selectControlsThreads

The number of parallel threads to use for selecting controls.

getDbExposureDataThreads

The number of parallel threads to use for fetching data on exposures for cases and controls.

createCaseControlDataThreads

The number of parallel threads to use for creating case and control data including exposure status indicators

fitCaseControlModelThreads

The number of parallel threads to use for fitting the models.

cvThreads

The number of parallel threads used for the cross-validation to determine the hyper-parameter when fitting the model.

Details

Run a list of analyses for the exposure-outcome-nesting cohorts of interest. This function will run all specified analyses against all hypotheses of interest, meaning that the total number of outcome models is `length(ccAnalysisList) * length(exposureOutcomeNestingCohortList)` (if all analyses specify an outcome model should be fitted). When you provide several analyses it will determine whether any of the analyses have anything in common, and will take advantage of this fact. For example, if we specify several analyses that only differ in the way the outcome model is fitted, then this function will extract the data and fit the propensity model only once, and re-use this in all the analysis.