Select matched controls per case

Select matched controls per case

selectControls(caseData, outcomeId, firstOutcomeOnly = TRUE,
  washoutPeriod = 180, controlsPerCase = 2, matchOnAge = TRUE,
  ageCaliper = 2, matchOnGender = TRUE, matchOnProvider = FALSE,
  matchOnCareSite = FALSE, matchOnVisitDate = FALSE,
  visitDateCaliper = 30, matchOnTimeInCohort = FALSE,
  daysInCohortCaliper = 30, minAge = NULL, maxAge = NULL,
  removedUnmatchedCases = TRUE, seed = NULL)

Arguments

caseData
An object of type caseData as generated using the getDbCaseData function.

outcomeId
The outcome ID of the cases for which we need to pick controls.

firstOutcomeOnly
Use the first outcome per person?

washoutPeriod
Minimum required numbers of days of observation for inclusion as either case or control.

controlsPerCase
Maximum number of controls to select per case.

matchOnAge
Match on age?

ageCaliper
Maximum difference (in years) in age when matching on age.

matchOnGender
Match on gender?

matchOnProvider
Match on provider (as specified in the person table)?

matchOnCareSite
Match on care site (as specified in the person table)?

matchOnVisitDate
Should the index date of the control be changed to the nearest visit date?

visitDateCaliper
Maximum difference (in days) between the index date and the visit date when matching on visit date.

matchOnTimeInCohort
Match on time in nesting cohort? When not using nesting, this is interpreted as time observed prior to index.

daysInCohortCaliper
Maximum difference (in days) in time in cohort.

minAge
Minimum age at which patient time will be included in the analysis. Note that information prior to the min age is still used to determine exposure status after the minimum age (e.g. when a prescription was started just prior to reaching the minimum age). Also, outcomes occurring before the minimum age is reached will be considered as prior outcomes when using first outcomes only. Age should be specified in years, but non-integer values are allowed. If not specified, no age restriction will be applied.

maxAge
Maximum age at which patient time will be included in the analysis. Age should be specified in years, but non-integer values are allowed. If not specified, no age restriction will be applied.

removedUnmatchedCases
Should cases with no matched controls be removed?

seed
The number generator seed. A null value sets seed via Sys.time.

Value

A data frame with these columns:

personId

The person ID

indexDate

The index date

isCase

Is the person a case or a control?

stratumId

The ID linking cases and controls in a matched set

Details

Select controls per case. Controls are matched on calendar time and the criteria defined in the arguments. Controls are randomly sampled to the required number.